perf: serve model JSON schemas as static files, remove Code bloat
Replace the 'API Schemas (Raw)' section's syntax-highlighted <Code>
blocks with static .json endpoints and a compact link UI.

The <Code> components were inflating 227 KB of raw JSON into ~1.95 MB
of syntax-highlighted HTML (8.6x ratio) via Expressive Code span wrapping.
On models with multiple API modes (sync/streaming/batch), the same
schema was rendered multiple times, pushing detail pages to ~3.1 MB.

Changes:
- Extract detectApiModes() from model-resolver.ts into model-schema.ts
  so it can be shared without pulling in Astro's getCollection
- Add src/pages/workers-ai/models/[...schema].json.ts: generates static
  JSON endpoints per model per schema (sync-input.json, sync-output.json,
  streaming-input.json, etc.) for all legacy models (~220 files)
- Add src/pages/ai/models/[...schema].json.ts: same for catalog models,
  handling multi-segment slugs (e.g. /ai/models/openai/gpt-5.4-mini/schema-input.json)
- Replace Code blocks in ModelDetailPage.astro with a 2-column card grid
  linking to the .json endpoints, each card having square open/download
  icon buttons

Expected result: ~3.1 MB -> ~1.15 MB on gemma detail page
